Symlink virtual machine into build directory; closes #145
$SMALLTALK_CI_VM cannot be set within run.sh, because it runs in a subshell.
We do not want it to run in the same shell. This way it is impossible to mess
with the user's shell.
Additionally, we do not know what VM to use until the image has been downloaded.
For this reason, the best solution is to symlink the VM just like the project
directory into the build directory.